+++ fitcrack/backend/src/src/api/fitcrack/endpoints/settings/responseModels.py
@@ -14,4 +14,5 @@ settings_model = api.model('Settings', {
     'bench_all': fields.Boolean(),
     'verify_hash_format': fields.Boolean(),
     'auto_add_hosts_to_running_jobs': fields.Boolean(),
+    'bench_runtime_limit': fields.Integer(),
 })