+++ fitcrack/backend/src/src/database/models.py
@@ -520,6 +520,7 @@ class FcSetting(Base):
     bench_all = Column(Integer, nullable=False, server_default=text("'1'"))
     verify_hash_format = Column(Integer, nullable=False, server_default=text("'1'"))
     auto_add_hosts_to_running_jobs = Column(Integer, nullable=False, server_default=text("'0'"))
+    bench_runtime_limit = Column(Integer, nullable=False, server_default=text("'30'"))
 
 class FcJobGraph(Base):
     __tablename__ = 'fc_job_graph'
@@ -710,9 +711,10 @@ class FcWorkunit(Base):
 
     @hybrid_property
     def remaining_time_str(self):
-        if not self.finished: # normal wu
-            if self.keyspace == 0: # benchmark wu has runtime limit 30 seconds
-                remaining_time = 0 if self.cracking_time > 30 else 30 - self.cracking_time
+        if not self.finished: # active wu
+            if self.keyspace == 0: # benchmark wu has runtime limit in seconds
+                settings = FcSetting.query.first()
+                remaining_time = 0 if self.cracking_time > settings.bench_runtime_limit else settings.bench_runtime_limit - self.cracking_time
                 return str(datetime.timedelta(seconds=math.floor(remaining_time)))
             elif self.keyspace > 0: # normal wu
                 return str(datetime.timedelta(seconds=math.floor(self.remaining_time)))