+++ server/src/source/Database/Job.cpp
@@ -83,6 +83,8 @@ CJob::CJob(DbMap &jobMap, CSqlLoader * sqlLoader)
 
         /** Temperature threshold */
         m_hwTempAbort = m_sqlLoader->getHWTempAbort();
+
+        m_benchRuntimeLimit = m_sqlLoader->getBenchRuntimeLimit();
     }
     catch(std::logic_error & error)
     {
@@ -357,6 +359,10 @@ uint64_t CJob::getFixedWorkunitSize() const
     return m_fixedWorkunitSize;
 }
 
+uint64_t CJob::getBenchRuntimeLimit() const
+{
+    return m_benchRuntimeLimit;
+}
 
 const std::string & CJob::getExtraHcArgs() const
 {